Refactored to move Article functionality to seperate document
